% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{cluster_haplotypes}
\alias{cluster_haplotypes}
\title{Cluster the most similar haplotypes together}
\usage{
cluster_haplotypes(sequences)
}
\arguments{
\item{sequences}{A character vector of sequences to cluster.}
}
\value{
A character vector with the same values as the input
  vector, but reordered to reflect similarity between strings.
}
\description{
This function takes a character vector of strings and uses a
greedy clustering algorithm based on distance between strings
to cluster them together.
}
\examples{
NULL
}
