% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_analysis.R
\name{compare_signals}
\alias{compare_signals}
\title{Check the overlap between newly-discovered signals and known signals}
\usage{
compare_signals(new_signals, known_signals)
}
\arguments{
\item{new_signals}{A GRanges object with the description of newly discovered signals}

\item{known_signals}{A data.frame object with the following metadata columns
(this data.frame will be coerced to a GRanges object internally: eventually
signals should be uniformly represented a GRanges throughout the package for
a cleaner interface)
(accessible through the \code{mcols} function :
\itemize{
  \item signal_key: the character key used to uniquely identify this signal
  \item n_snps: the number of SNPs associated with that signal in the original study
  \item log_pvalue: the maximum -log10(p-value) observed for all associated SNPs
  \item max_pos: the position along the reference where the most significant p-value occurs
  \item locus: the name by which the locus is typically referred to
  \item glyma_name: the standardized name by which the gene is known
  \item common_name: the name by which the gene is commonly known
  \item size: the length of the gene
}}
}
\value{
a list of two GRanges object. The first element of the list
  is a GRanges similar to new_signals but with added metadata columns
  with information about the matches that were found in known_signals.
  Values are NA where no match was found (or the value was already NA
  in known_signals). The second element of the list is similar except
  that it gives the information on known_signals and the new_signals
  that overlapped them.
}
\description{
Check the overlap between newly-discovered signals and known signals
}
\examples{
NULL
}
