% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_analysis.R
\name{extract_genes}
\alias{extract_genes}
\title{Get the genes associated with a set of GWAS signals}
\usage{
extract_genes(signals, genes)
}
\arguments{
\item{signals}{A GRanges object of significant GWAS signals, as output
by the function \code{\link{extract_signals}}.}

\item{genes}{A GRanges object describing the set of genes in the organism
of interest, as returned by the function \code{\link[GenomicFeatures]{genes}}}
}
\value{
To be completed
}
\description{
Get the genes associated with a set of GWAS signals
}
\examples{
NULL
}
