% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_querying.R
\name{extract_reads}
\alias{extract_reads}
\title{extract_reads}
\usage{
extract_reads(kmer_sequence, bam_file, samtools_path, output)
}
\arguments{
\item{kmer_sequence}{To be completed}

\item{bam_file}{To be completed}

\item{samtools_path}{To be completed}

\item{output}{To be completed}
}
\value{
To be completed
}
\description{
To be completed
}
\examples{
NULL
}
