% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_analysis.R
\name{extract_signals}
\alias{extract_signals}
\title{Extract signal ranges from GWAS results}
\usage{
extract_signals(markers, threshold, distance = 10^5)
}
\arguments{
\item{markers}{A GRanges object of formatted GWAS association
data as generated by \code{\link{format_kmer_gwas}} or
\code{\link{format_gapit_data}}.}

\item{threshold}{Numeric. The minimum -log10(p-value) to consider
as a significant association}

\item{distance}{Numeric. The maximum distance (in base pairs) that
two ranges must be for them to be merged into the same signal.}
}
\value{
A GenomicRanges object with the grouped signals. The metadata
  columns "MAPQ" and "log10p" in the output GRanges represent the 
  maximum values across all markers in the signal.
}
\description{
This function takes a GRanges object of GWAS association data
formatted with either \code{\link{format_kmer_gwas}} or
\code{\link{format_gapit_data}} and groups contiguous
significant associations into the same GRanges
object (given some criteria) so that it can be used
to query the genome e.g. for overlapping genes or variants.
}
\examples{
NULL
}
