% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill the deleted positions with dashes}
\usage{
fill_gaps(hapdata)
}
\arguments{
\item{hapdata}{A list of data.frames containing data about
haplotypes to be plotted, possibly with missing positions
due to alignment gaps.}
}
\value{
A list of data.frames similar in format to hapdata,
  in which positions that correspond to gaps have been filled.
}
\description{
This function takes a list of data.frames that have been
adjusted for position by taking alignment gaps into account,
and fills those positions with gaps by setting the nucleotide
to a dash ('-') and the -log10(p-value) to \code{NA}. This
function is not exported for direct use, but is instead called
internally by \code{\link{adjust_gaps}} to simplify its code.
}
\examples{
NULL
}
