% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{format_gapit_gwas}
\alias{format_gapit_gwas}
\title{Format GAPIT GWAS results for generating Manhattan plots}
\usage{
format_gapit_gwas(gapit_df, fai_file, chromosomes, pattern = NULL)
}
\arguments{
\item{fai_file}{A character of length one. A .fai index file that
will be queried with the \code{\link{parse_fai}} function to
query information about the reference genome and translate
positions on specific input sequences to positions along the
whole reference.}

\item{chromosomes}{a character vector of the same length
as the number of unique values in the Chromosome column
of the GAPIT output. The integer chromosome values in the
results returned by GAPIT will be used to index into that
character vector to translate integer values into chromosome
names.}

\item{pattern}{A character of length one. A regular expression
used to match the name of the chromosome or scaffolds in the
.fai index file. Matching sequences will be kept. If NULL
(the default), all sequences are kept.}

\item{sam_df}{A data.frame of GAPIT analysis results as returned
by \code{\link{read_gapit}}.}
}
\value{
A data.frame that contains the following standard columns
  and can be used for generating a Manhattan plot using the 
  function manhattan_plot.
  \itemize{
    \item manhattan_chrom Character. The name of the chromosome
    \item manhattan_rpos  Numeric. The position along the chromsome
      expressed as a cumulative position along the reference
    \item manhattan_cpos Numeric. The position on the reference
      expressed in terms of position along a particular chromosome
    \item mahattan_log10p The p-value used for plotting, obtained by
      computing -log10 of the p-value.
    \item manhattan_even Logical. \code{TRUE} if the chromosome should be
      considered even-valued, \code{FALSE} otherwise. Used to
      distinguish the chromosome that are next to each other by color.
  }

  These data columns are guaranted to be the first ones, in that order,
  in the output data.frame. Any other columns that were present in the
  input are appended to the output data.frame as is.
}
\description{
This function takes a data.frame returned by the \code{\link{read_gapit}} function
and returns a standard data.frame for Manhattan plotting
}
\examples{
NULL
}
