% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{format_haplotypes}
\alias{format_haplotypes}
\title{Format haplotypes for plotting using k-mer overlap information}
\usage{
format_haplotypes(haplotypes, overlaps)
}
\arguments{
\item{haplotypes}{To complete}

\item{overlaps}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
