% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{format_kmer_gwas}
\alias{format_kmer_gwas}
\title{Format k-mer GWAS results for downstream analyses}
\usage{
format_kmer_gwas(sam_df, ref_fasta, pattern = NULL, min_mapq = 0)
}
\arguments{
\item{sam_df}{A data.frame of k-mer matches to alignments annotated
with p-values, as returned by the function \code{\link{add_pvalues}}}

\item{ref_fasta}{A character of length one. A fasta file that
will be queried with the \code{\link[Rsamtools]{scanFaIndex}} function to
query information about the reference genome and set the seqinfo fields
of the output GRanges object.}

\item{pattern}{A character of length one. A regular expression
used to match seqlevels of the chromsomes or scaffolds in the
fasta index. Matching sequences will be kept. If NULL
(the default), all sequences are kept. In the current implementation,
an error will be thrown if changing the seqlevels results in matches
being pruned.}

\item{min_mapq}{A numeric of length one. The minimum mapping quality
of the alignment on which the k-mer was found for it to be kept.
The default value (0) implies no filtering.}
}
\value{
A GRanges object that contains the information on matching
  k-mers and can be used for downstream analyses including the
  function \code{\link{manhattan_plot}}. Must contain a field
  called log10_p that stores the -log10(p-value) of the k-mer.
}
\description{
This function takes a data.frame returned by the \code{\link{add_pvalues}}
function and returns a GRanges object suitable for downstream analyses 
and plotting.
}
\examples{
NULL
}
