% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{gapit_manhattan}
\alias{gapit_manhattan}
\title{gapit_manhattan}
\usage{
gapit_manhattan(
  formatted_gapit,
  fai_file,
  pattern = NULL,
  signals = NULL,
  single_panel = TRUE,
  threshold = 5
)
}
\arguments{
\item{formatted_gapit}{To be completed}

\item{fai_file}{To be completed}

\item{pattern}{To be completed}

\item{signals}{To be completed}

\item{single_panel}{To be completed}

\item{threshold}{To be completed}
}
\value{
To be completed
}
\description{
gapit_manhattan
}
\examples{
NULL
}
