% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_analysis.R
\name{gapit_vcf}
\alias{gapit_vcf}
\title{Launch a MLM GWAS analysis on a VCF file read with VariantAnnotation::readVcf using GAPIT}
\usage{
gapit_vcf(vcf, kinship, pca, phenodata, trait, id_column, tmproot, ...)
}
\arguments{
\item{kinship}{A kinship data.frame satisfying the input requirements of GAPIT.}

\item{pca}{A pca data.frame satisfying the input requirements of GAPIT.}

\item{phenodata}{A character, the name of a file containing semicolon-separated
phenotypic data}

\item{trait}{A character, the name of the trait to analyze. Should correspond
to the name of a column in the phenotypic data.}

\item{id_column}{A character, the name of the column containing the sample IDs
in the phenotypic data. These sample IDs should correspond to those used in the
VCF file.}

\item{tmproot}{A character. The temporary directory to use as a root for creating
temporary storage.}

\item{...}{Other arguments to be passed to GAPIT}

\item{A}{CollapsedVcf object read by \code{\link[VariantAnnotation]{readVcf}}.}
}
\value{
A GRanges object containing the results of the GAPIT analysis.
}
\description{
This function is used as a convenience to quickly launch a GAPIT analysis based on
data read with \code{\link[VariantAnnotation]{readVcf}}.
}
\examples{
NULL
}
