% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{get_haplotypes}
\alias{get_haplotypes}
\title{Extract the haplotypes found in a set of sequences and filter based on frequency}
\usage{
get_haplotypes(sequences, min_frequency)
}
\arguments{
\item{sequences}{To complete}

\item{min_frequency}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
