% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_querying.R
\name{get_samples}
\alias{get_samples}
\title{get_samples}
\usage{
get_samples(pav_table, kmer_list, index)
}
\arguments{
\item{pav_table}{To be completed}

\item{kmer_list}{To be completed}

\item{index}{To be completed}
}
\value{
To be completed
}
\description{
To be completed
}
\examples{
NULL
}
