% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Select colors for a discrete scale as in ggplot2}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{A numeric value. The number of colors to generate.}
}
\value{
A vector of colors to use for plotting.
}
\description{
This code is taken from https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
It is meant to be used internally by the package and is therefore not exported.
}
\examples{
NULL
}
