% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{gg_hue}
\alias{gg_hue}
\title{Generating a vector of colors comparable to those used by ggplot2}
\usage{
gg_hue(n)
}
\arguments{
\item{n}{An integer of length one. The number of colors to generate}
}
\value{
A vector of colors to be used for plotting
}
\description{
Generating a vector of colors comparable to those used by ggplot2
}
\examples{
NULL
}
