% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_functions.R
\name{kmer_ld}
\alias{kmer_ld}
\title{Compute the pairwise LD between k-mers based on their presence/absence}
\usage{
kmer_ld(pav, kmers)
}
\arguments{
\item{pav}{A matrix with presence/absence of a set of k-mers}

\item{kmers}{A character vector of the k-mer sequences}
}
\value{
A numeric matrix with the pairwise LD values computed between
  all k-mers.
}
\description{
This function has been obsoleted by a much faster C implementation
}
\examples{
NULL
}
