% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{kmer_manhattan}
\alias{kmer_manhattan}
\title{kmer_manhattan}
\usage{
kmer_manhattan(formatted_df, fai_file, pattern = NULL, signals = NULL)
}
\arguments{
\item{formatted_df}{To be completed}

\item{fai_file}{To be completed}

\item{pattern}{To be completed}

\item{signals}{To be completed}
}
\value{
To be completed
}
\description{
kmer_manhattan
}
\examples{
NULL
}
