% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_functions.R
\name{ld_plot}
\alias{ld_plot}
\title{Plot an LD matrix using grid functions}
\usage{
ld_plot(
  ld_matrix,
  kmer_positions,
  top_legend = TRUE,
  ylabels = FALSE,
  ylabel_pattern = NULL,
  fontsize = 8
)
}
\arguments{
\item{ld_matrix}{A matrix of pairwise linkage disequilibrium values observed
among a set of k-mers.}

\item{kmer_positions}{A Granges object storing the genomic positions and
p-values associated with a set of significant k-mers. Must contain
a 'log10p' columns containing the -log10(p-value) associated with each k-mer.
All k-mers represented in the ld_matrix object must also be presented within
kmer_positions.}

\item{top_legend}{A logical. Whether or not to plot the color scale for the
chromosomes at the top of the plot. Defaults to TRUE.}

\item{ylabels}{A logical. If TRUE, labels are printed on the y-axis at the mid-position
for each reference sequence to identify them in the plot. Setting this option to
TRUE only makes sense when k-mers are sorted by genomic position, therefore it is
FALSE by default.}

\item{ylabel_pattern}{A regular expression to use to keep only the labels matching
that pattern. This option is only relevant if ylabels is set to TRUE.}

\item{fontsize}{A numeric. The font size to use for the y-labels and color scales.}
}
\value{
\code{NULL}, invisibly. This function is invoked for its plotting side effect.
}
\description{
This function uses grid functionality to plot an LD matrix.
LD values are represented using different shades from yellow
to red as implemented by the RColorBrewer "YlOrRd" scale.
The chromosome or scaffold to which a k-mer belong is represented
by colored rectangles plotted in the left outer margin of the plot.
Similarly, the -log10(p-values) are represented using the "Blues"
color scale of the package RColorBrewer in the bottom outer
margin of the plot. The y-axis of the generated plot is inverted
with respect to the input matrix, i.e., the first row of the matrix
is found at the very bottom of the plot and the last row of the matrix
is found at the very top of the plot.
}
\examples{
NULL
}
