% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_functions.R
\name{ld_sort}
\alias{ld_sort}
\title{Re-arrange the samples in an LD matrix according to user-specified criteria}
\usage{
ld_sort(
  ld_matrix,
  kmer_data,
  sort_param = c("position", "pvalue"),
  positions = NULL
)
}
\arguments{
\item{ld_matrix}{A square numeric matrix with each cell indicating the pairwise
linkage disequilibrium between two k-mers. The rows and columns of the matrix
must be named after the k-mers.}

\item{kmer_data}{A GRanges object indicating the genomic location of significantly
associated k-mers. This object must have its elements named after the k-mer
sequence (for subsetting purposes) and have a metadata column named "log10p"
for sorting based on p-values.}

\item{sort_param}{A parameter to use for sorting the LD matrix. At the moment
two values are supported:
\itemize{
  \item "position": k-mers are sorted according to their genomic position 
  \item "pvalue": k-mers are sorted from the most significant to least significant p-value
}}

\item{positions}{A GRanges object used to restrict the LD matrix to the k-mers overlapping
a given set of genomic positions. If \code{NULL} (the default), then no
subsetting is performed.}
}
\value{
An LD matrix that has been reordered according to the choice of sort_param,
  and possibly subset based on the positions parameter.
}
\description{
This function takes an LD matrix as input, a dataset of k-mer positions (with
associated p-values), and a sorting parameter, and returns the same LD matrix
with the order of the samples re-arranged according to the sorting parameter.
}
\examples{
NULL
}
