% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{manhattan_plot}
\usage{
manhattan_plot(
  formatted_data,
  gwas_type,
  fai_file,
  pattern = NULL,
  signals = NULL,
  extent = FALSE,
  single_panel = TRUE,
  threshold = 5,
  min_log10p = 0
)
}
\arguments{
\item{formatted_data}{A data.frame formatted with either \code{\link{format_kmer_gwas}}
or \code{\link{format_gapit_gwas}} depending on the GWAS type. These functions
return data.frames containing standard columns in addition to GWAS type-specific
columns, therefore allowing a common interface for all GWAS types.}

\item{gwas_type}{A character of length one. Only two options are supported at
this time, either \code{"kmer"} for k-mer GWAS as implemented by Voichek
and Weigel (2020) or \code{"gapit"} for the output of GAPIT.}

\item{fai_file}{A character of length one. The name of the .fai index file
to extract the reference genome data from.}

\item{pattern}{A character of length one. A regular expression to use
for selecting what reference sequences to keep. If \code{NULL}
(the default), all sequences are used. The function will force all
references sequences kept to be plotted even if no data point or
signal is observed on it.}

\item{signals}{A data.frame with the coordinates of GWAS signals to 
show on the plot. If \code{NULL} (the default) or \code{nrow(signals)} < 0,
then no signals are plotted. If supplied, the data.frame must minimally
contain the following columns:
\itemize{
  \item  first_chrom: the chromosome where the leftmost position is found
  \item  first_pos: the coordinate on first_chrom where the leftmost significant marker occurs
  \item  last_chrom: the chromosome where the rightmost position is found
  \item  last_pos: the coordinate on last_chrom where the rightmost significant marker occurs
  \item  max_chrom: the chromosome where the marker with the most significant p-value occurds
  \item  max_pos: the coordinate on max_chrom where the most significant marker occurs
}
Supplying this data.frame in itself only results in a vertical line being
drawn at the location of each signal. If, in addition, \code{extent = TRUE},
then the extent of the signal (from the first to the last position along
the chromosome) is also plotted as a rectangle.}

\item{extent}{Logical. Whether or not to plot the extent of the signals
as a rectangle. Can only be \code{TRUE} if signals is not \code{NULL}.}

\item{single_panel}{Logical. Whether all chromosomes should be plotted in
a single panel (the default) or \code{\link[ggplot]{facet_wrap}} should
be used to plot each chromosome in a separate panel.}

\item{threshold}{Numeric. A single value indicating where the -log10(p-value)
threshold line should be plotted.}

\item{min_log10p}{Numeric. The minimum -log10(p-value) to display on the plot
can be used to filter out some non-significant markers for plotting speedup.}
}
\value{
A ggplot object describing a Manhattan plot with the specified parameters.
}
\description{
manhattan_plot
}
\examples{
NULL
}
