% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{manhattanGrob}
\alias{manhattanGrob}
\title{A function that returns a graphical object (grob) representing a manhattan plot}
\usage{
manhattanGrob(
  gwas_results,
  threshold = NULL,
  min_log10p = 0,
  numeric_chrom = FALSE,
  yexpand = c(0.03, 0.08),
  xexpand = c(0.03, 0.03),
  cex.points = 0.5,
  cex.lab = 1,
  margins = c(5.1, 4.1, 4.1, 2.1)
)
}
\arguments{
\item{gwas_results}{A GRanges object formatted with either \code{\link{format_kmer_gwas}}
or \code{\link{format_gapit_gwas}}. These functions
return GRanges containing standard columns and seqinfo in addition to GWAS type-specific
columns, therefore allowing a common interface for all GWAS types.}

\item{threshold}{Numeric. A single value indicating where the -log10(p-value)
threshold line should be plotted.}

\item{min_log10p}{Numeric. The minimum -log10(p-value) to display on the plot
can be used to filter out some non-significant markers for plotting speedup.}

\item{numeric_chrom}{Logical. Whether chromosome names should be stripped from
their alphabetical component and converted to numeric values (default: FALSE).}

\item{yexpand}{A numeric of length two. A fractional value relative to the
range of the y-axis used to expand the y-scale on either side. The first
value is the expansion factor at the bottom of the scale, and the second
value is the expansion factor at the top of the scale.}

\item{xexpand}{Same as yexpand but for the x-scale.}

\item{cex.points}{A numeric value. The expansion factor for the points.}

\item{cex.lab}{A numeric value. The expansion factor for the axis labels.}

\item{margins}{The margins used for the plotting viewport. Default value
is c(5.1, 4.1, 4.1, 2.1) (the default for \code{\link[grid]{plotViewport}})}
}
\value{
A gTree describing a Manhattan plot with the specified parameters.
}
\description{
A function that returns a graphical object (grob) representing a manhattan plot
}
\examples{
NULL
}
