% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{map_color}
\alias{map_color}
\title{Map a set of numeric values onto a color palette}
\usage{
map_color(values, pal, n_colors, scale_buffer)
}
\arguments{
\item{values}{To complete}

\item{pal}{To complete}

\item{n_colors}{To complete}

\item{scale_buffer}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
