% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{match_kmers}
\alias{match_kmers}
\title{A function that matches a set of significant k-mers to positions on haplotypes}
\usage{
match_kmers(haplotypes, kmers, kmer_length)
}
\arguments{
\item{haplotypes}{To complete}

\item{kmers}{To complete}

\item{kmer_length}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
