% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{nucdiff}
\alias{nucdiff}
\title{Find positions that differ between two sequences}
\usage{
nucdiff(hapdata)
}
\arguments{
\item{hapdata}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
