% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{parse_fai}
\alias{parse_fai}
\title{Parse a .fai fasta index file}
\usage{
parse_fai(fai_file, pattern = NULL)
}
\arguments{
\item{fai_file}{Character of length one. The path to the .fai index file.}

\item{pattern}{Character of length one. A regular expression used to select
the chromosomes or scaffolds to keep. If NULL, all entries in the .fai
index are kept}
}
\value{
A list of length three with the following elements:
  \itemize{
    \item lengths: Numeric. The length of each sequence
    \item start: Numeric. The (cumulative) start position of that 
                 sequence along the genome
    \item label_pos: Numeric. The position to use for plotting chromosome
                     labels in the Manhattan plot.
  }
  All of the vectors in the list are named according to the chromosome/scaffold
}
\description{
This function takes a .fai index and returns some data that is useful
for generating Manhattan plots.
}
\examples{
NULL
}
