% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_querying.R
\name{parse_sam_data}
\alias{parse_sam_data}
\title{parse_sam_data}
\usage{
parse_sam_data(sam_paths)
}
\arguments{
\item{sam_paths}{To be completed}
}
\value{
To be completed
}
\description{
To be completed
}
\examples{
NULL
}
