% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_transcript}
\alias{plot_transcript}
\title{Plot a transcript using grid functions}
\usage{
plot_transcript(txdb, gene_name, tx_name)
}
\arguments{
\item{txdb}{a TxDb object containing the information on transcripts
in the genome of interest.}

\item{gene_name}{A character. The name of the gene in the TxDb object
that the transcript is a part of.}

\item{tx_name}{A character. The name of the transcript in the TxDb object
to be plotted.}
}
\value{
NULL, invisibly. The function is invoked for its side
  effect of plotting the transcript.
}
\description{
This function plots a transcript in a TxDb object using grid
functions. This assumes that tht plotting occurs in a viewport
with native coordinates already suitable for plotting that
transcript.
}
\examples{
NULL
}
