% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvalue_plot}
\alias{pvalue_plot}
\title{Plot the p-values of a GWAS analysis at a locus using grid functions}
\usage{
pvalue_plot(gwas_results, interval)
}
\arguments{
\item{gwas_results}{A data.frame of GWAS results including minimally
the following columns for plotting: manhattan_cpos, manhattan_chrom,
and manhattan_log10p. Can also be directly provided as a GRanges object.}

\item{interval}{A GRanges object used to subset the plotting to a given
region.}
}
\value{
NULL, invisibly. This function is called for its plotting
  side-effects.
}
\description{
This function takes a data.frame of formatted GWAS results and plots
the p-values in a given interval.
}
\examples{
NULL
}
