% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{read_consensus}
\alias{read_consensus}
\title{Reading the consensus sequences obtained at a locus for a set of samples}
\usage{
read_consensus(input_fasta)
}
\arguments{
\item{input_fasta}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
