% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file into a named character vector}
\usage{
read_fasta(input_fasta)
}
\arguments{
\item{input_fasta}{A character. The name of the fasta file to
read the input from.}
}
\value{
A character vector with as many elements as there are
  sequences. The name of each element is the name of the sequence
  (the sequence of characters that follows ">").
}
\description{
This function reads a fasta file into a named character vector.
Its functionality is limited as it is required that all sequences
be on a single line as the function assumes one line per sequence.
}
\examples{
NULL
}
