% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{read_gapit}
\alias{read_gapit}
\title{Read the output of GAPIT}
\usage{
read_gapit(filename)
}
\arguments{
\item{filename}{Character of length one. The name of the .csv file
containing the results of the GAPIT analysis.}
}
\value{
a data.frame containing the values read from the file
}
\description{
This function reads the output of an analysis with GAPIT.
Only relevant fields (Chromosome, Position, P.value, FDR_Adjusted_P-values)
are kept to increase reading efficiency.
}
\examples{
NULL
}
