% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_functions.R
\name{subsample_kmers}
\alias{subsample_kmers}
\title{Subsample from a set of significant k-mers prior to computing LD}
\usage{
subsample_kmers(kmer_data, nkmers, npvalue)
}
\arguments{
\item{kmer_data}{A GRanges object containing the genomic positions of 
significantly associated k-mers. Must containg a "log10p" column
to sample k-mers based on their p-values, and a "kmer_canon" column
containing the canonized sequence of the k-mer matching at that position.}

\item{nkmers}{A numeric indicating the total number of k-mers to subsample.}

\item{npvalue}{A numeric indicating the number of k-mers to sample from
the top p-values. The npvalue most associated k-mers will be sampled prior
to random sampling of the remaining k-mers.}
}
\value{
A subset of the input kmer_data object containing the only the
  subsampled k-mers.
}
\description{
Computing the pairwise LD between all k-mers in a set is computationally
expensive as the number of comparisons made is the square of the number
of input k-mers. Therefore, this function allows for subsampling from
a set of input k-mers such that the most significant k-mers are represented
while sampling from the whole genome and not just from the most significant
regions. This is acheived from selecting a fixed number of top significant
k-mers (the npvalue parameter) and sampling from the remaining k-mers
with a probability that is inversely proportional to the number of k-mers
found on a given sequence. This ensures that k-mers are sampled from
various regions of the genome.
}
\examples{
NULL
}
