% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{transcriptGrob}
\alias{transcriptGrob}
\title{Generate a grob representing a transcript using grid functions}
\usage{
transcriptGrob(tx_gene, tx_exons, tx_cds)
}
\arguments{
\item{tx_gene}{A GRanges object representing the coordinates of
of a single gene to which the transcript belongs.}

\item{tx_exons}{A GRanges object representing the exons belonging
to a single transcript.}

\item{tx_name}{A GRanges object representing the coding sequences
belonging to a single transcript.}
}
\value{
A gList grob object representing a single transcript.
}
\description{
This function generates a grob of a transcript model using grid functions.
}
\details{
The function assumes that that the plotting will occur in a viewport
with x-axis native coordinates already suitable for plotting that
transcript. y-axis coordinates do not matter and "npc" coordinates
are used for centering the transcript models vertically in the
viewport.

The function will be typically called by the wrapper function
\code{\link{plot_transcripts}} that will take care of organizing
the viewports and coordinate scales for plotting all the transcripts
of a given gene model.
}
\examples{
NULL
}
