% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_analysis.R
\name{vcf_to_gapit}
\alias{vcf_to_gapit}
\title{Convert VCF records read with VariantAnnotation::readVCF into a format usable by GAPIT}
\usage{
vcf_to_gapit(x)
}
\arguments{
\item{x}{A CollapsedVCF object as read by \code{\link[VariantAnotation]{readVcf}}}
}
\value{
A data.frame object representing data in HapMap format, suitable for
  input as the genotype argument of \code{\link[GAPIT3]{GAPIT}}
}
\description{
Convert VCF records read with VariantAnnotation::readVCF into a format usable by GAPIT
}
\examples{
NULL
}
