% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{adjust_gaps}
\alias{adjust_gaps}
\title{Adjust plotting position according to the presence of gaps}
\usage{
adjust_gaps(hapdata, gaps)
}
\arguments{
\item{hapdata}{To complete}

\item{gaps}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
