% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{format_gapit_gwas}
\alias{format_gapit_gwas}
\title{Format GAPIT GWAS results for generating Manhattan plots}
\usage{
format_gapit_gwas(filename, ref_fasta, chromosomes, pattern = NULL)
}
\arguments{
\item{filename}{Character of length one. The name of the .csv file
containing the results of the GAPIT analysis.}

\item{ref_fasta}{A character of length one. A fasta file that
will be queried with the \code{\link[Rsamtools]{scanFaIndex}} function to
query information about the reference genome and set the seqinfo fields
of the output GRanges object.}

\item{chromosomes}{a character vector of the same length
as the number of unique values in the Chromosome column
of the GAPIT output. The integer chromosome values in the
results returned by GAPIT will be used to index into that
character vector to translate integer values into chromosome
names.}

\item{pattern}{A character of length one. A regular expression
used to match seqlevels of the chromsomes or scaffolds in the
fasta index. Matching sequences will be kept. If NULL
(the default), all sequences are kept. In the current implementation,
an error will be thrown if changing the seqlevels results in matches
being pruned.}
}
\value{
A GRanges object that contains the information on associated
  markers and can be used for downstream analyses including the
  function \code{\link{manhattan_plot}}. Must contain a field
  called log10_p that stores the -log10(p-value) of the marker.
}
\description{
This function takes a CSV file of GAPIT GWAS results and formats
them as a GRanges object for processing with other functions, including
\code{\link{manhattan_plot}}.
}
\details{
Only relevant fields (Chromosome, Position, P.value, FDR_Adjusted_P-values)
are kept to increase reading efficiency.
}
\examples{
NULL
}
