% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_parsing.R
\name{format_kmer_gwas}
\alias{format_kmer_gwas}
\title{Format k-mer GWAS results for downstream analyses}
\usage{
format_kmer_gwas(input_bam, ref_fasta, pattern = NULL, min_mapq = 0)
}
\arguments{
\item{input_bam}{A character. The path to an indexed .bam file that
contains KM and PV tags indicating the matching k-mers and their respective
p-values}

\item{ref_fasta}{A character of length one. A fasta file that
will be queried with the \code{\link[Rsamtools]{scanFaIndex}} function to
get information about the reference genome and set the seqinfo fields
of the output GRanges object.}

\item{pattern}{A character of length one. A regular expression
used to match seqlevels of the chromsomes or scaffolds in the
fasta index. Matching sequences will be kept. If NULL
(the default), all sequences are kept. In the current implementation,
an error will be thrown if changing the seqlevels results in matches
being pruned. NO LONGER TRUE}

\item{min_mapq}{A numeric of length one. The minimum mapping quality
of the alignment on which the k-mer was found for it to be kept.
The default value (0) implies no filtering.}
}
\value{
A GRanges object that contains the information on matching
  k-mers and can be used for downstream analyses including the
  function \code{\link{manhattan_plot}}. Must contain a field
  called log10_p that stores the -log10(p-value) of the k-mer.
}
\description{
This function takes a sorted and indexed .bam file of reads that matched 
significant k-mers as obtained from the katcher program and annotated
with p-values by the add_pvalues program. It returns a GRanges object
formatted for downstream analyses, including manhattan plotting.
}
\examples{
NULL
}
