% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_analysis.R
\name{gene_distance}
\alias{gene_distance}
\title{Get the signed distance between a gene and markers in a signal}
\usage{
gene_distance(signals, markers, gene)
}
\arguments{
\item{signals}{A GRanges object of GWAS signals to check for potential
overlaps with the genes.}

\item{markers}{A GRanges object of markers used in GWAS.}

\item{gene}{A GRanges object containing the positions of a gene of
interest. In the current implementation it is an error to pass
a GRanges object with more than one range.}
}
\value{
A GRanges object of the same form as the markers GRanges
  but with an added column "distance" showing the (signed) distance
  from each marker to the gene of interest. Only the markers in the
  signal that overlaps the genes are returned. If none of the signals
  or no markers are located in the interval of interest, then an empty
  GRanges object is returned.
}
\description{
This function takes a set of signals found by a GWAS analysis
and extracts the one (if any) that overlaps a gene of interest.
If there is one such signal, then the distance from all the
markers in that signal and the gene of interest is computed and
returned.
}
\examples{
NULL
}
