% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{grid.colorscale}
\alias{grid.colorscale}
\title{Plot the color scale used in a haplotype plot}
\usage{
grid.colorscale(breaks, base_palette)
}
\arguments{
\item{breaks}{A numeric vector of breaks used for the color scale.
There must be one more break than the number of colors}

\item{base_palette}{A character vector of colors used in mapping
the numeric values onto the color scale.}
}
\value{
\code{NULL}, invisibly. This function is invoked for its
  plotting side-effects.
}
\description{
This function takes a set of numeric breaks and a set of colors
and uses this information to produce a graphical scale linking
those breaks to the colors used in a plot. This function is not
exported from the package and is therefore meant to be used only
internally by the package.
}
\examples{
NULL
}
