% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{grid.colorscale}
\alias{grid.colorscale}
\title{Plot the color scale used in a haplotype plot}
\usage{
grid.colorscale(breaks, base_palette)
}
\arguments{
\item{breaks}{To complete}

\item{base_palette}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
