% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{grid.haplotypes}
\alias{grid.haplotypes}
\title{Plotting haplotypes}
\usage{
grid.haplotypes(plotting_data, difflist, fontsize = 8)
}
\arguments{
\item{plotting_data}{To complete}

\item{difflist}{To complete}

\item{fontsize}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
