% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{grid.phenotable}
\alias{grid.phenotable}
\title{Plot a table of phenotypes versus haplotypes}
\usage{
grid.phenotable(haplotype_data)
}
\arguments{
\item{haplotype_data}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
