% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{link_phenotypes}
\alias{link_phenotypes}
\title{A function that links haplotypes and the corresponding observed phenotypes}
\usage{
link_phenotypes(sequences, haplotypes, phenotypes, id_column, phenotype_column)
}
\arguments{
\item{sequences}{To complete}

\item{haplotypes}{To complete}

\item{phenotypes}{To complete}

\item{id_column}{To complete}

\item{phenotype_column}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
