% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{mafft_align}
\alias{mafft_align}
\title{Align a set of haplotypes using mafft}
\usage{
mafft_align(fasta_path, haplotypes, mafft_path, mafft_options)
}
\arguments{
\item{fasta_path}{To complete}

\item{haplotypes}{To complete}

\item{mafft_path}{To complete}

\item{mafft_options}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
