% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{map_color}
\alias{map_color}
\title{Map a set of numeric values onto a color palette}
\usage{
map_color(values, max, min = 0, pal = "YlOrRd", n.colors = 9)
}
\arguments{
\item{values}{To complete}

\item{max}{To complete}

\item{min}{To complete}

\item{pal}{To complete}

\item{n.colors}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
