% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{map_color}
\alias{map_color}
\title{Map a set of numeric values onto a color palette}
\usage{
map_color(values, pal, n_colors, scale_extend, n_colors_remove = 0)
}
\arguments{
\item{values}{A numeric vector of values to be mapped on the color scale.
\code{NA} values are accepted and mapped to the "black" color.}

\item{pal}{A character. The palette to use for to -log10(p-values) color scale.
must correspond to a palette in the RColorBrewer package.}

\item{n_colors}{A numeric. The number of colors to use for the mapping p-values
onto a color scale. The maximum value allowed may differ depending on the
particular palette used. This represents the number of colors in the base
palette prior to applying the n_colors_remove option.}

\item{scale_extend}{A numeric. The number of units (of -log10(p-value)) to extend
the limits of the color scale by at both ends to make sure that extreme values
are mapped onto the scale.}

\item{n_colors_remove}{A numeric. The number of colors to remove from the beginning
of the color scale, to remove paler hues. If this option is used, the length of the
final palette will be n_colors - n_colors_remove.}
}
\value{
A list of length three comprising the following elements:
  \itemize{
    \item mapped_colors: a character vector of colors mapped from the values input vector
    \item breaks: a numeric vector of break positions used for mapping
    \item base_palette: the colors that were used for mapping the numeric values
  }
}
\description{
This function takes a set of numeric values and maps them on a color palette
provided by the RColorBrewer package. \code{NA} values are mapped to black
as a default color for missing values.
}
\examples{
NULL
}
