% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{nucdiff}
\alias{nucdiff}
\title{Find positions that differ between two sequences}
\usage{
nucdiff(hapdata)
}
\arguments{
\item{hapdata}{A list of data.frames prepared for plotting,
as returned by \code{\link{adjust_gaps}}.}
}
\value{
A list of numeric vectors indicating the positions in
  hapdata where top consecutive haplotypes differ. The length
  of the returned list is therefore length(hapdata) - 1.
}
\description{
This function takes as input a list of haplotypes that have been
prepared for plotting and identifies the positions. The positions
of the nucleotides must have been prepared with \code{\link{adjust_gaps}}
such that there is a one-to-one correspondence between positions.
Gaps are not flagged as differing between sequences.
}
\examples{
NULL
}
