% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_transcripts}
\alias{plot_transcripts}
\title{Plot all the possible transcripts for a gene using grid functions}
\usage{
plot_transcripts(txdb, gene_name)
}
\arguments{
\item{txdb}{A TxDb object containing the genes and transcripts
of a genome of interest.}

\item{gene_name}{A character. The name of the gene whose transcripts
are to be plotted.}
}
\value{
NULL, invisibly. The function is called for its plotting
  side-effect.
}
\description{
This function calls plot_transcript on all individual transcripts
of a gene and plots each separately in a row of a viewport grid.
}
\examples{
NULL
}
