% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvalue_plot}
\alias{pvalue_plot}
\title{Plot the p-values of a GWAS analysis at a locus using grid functions}
\usage{
pvalue_plot(
  gwas_results,
  interval,
  pvalue_margins = c(5.1, 4.1, 4.1, 2.1),
  yexpand = c(0, 0)
)
}
\arguments{
\item{gwas_results}{A GRanges object or data.frame of GWAS results.
If a data.frame, must include the columns manhattan_chrom and
manhattan_cpos for coercion to a GRanges. The metadata column
log10p must be present as it will be used to plot the
p-values on the y-axis.}

\item{interval}{A GRanges object used to subset the plotting to a given
region.}

\item{pvalue_margins}{A numeric of length 4 or that can be recycled to that
length. The margins used by the function \code{\link[grid]{plotViewport}}
to set the margins around the plotting region. By default it is set
to c(5.1, 4.1, 4.1, 2.1).}

\item{yexpand}{A numeric of length 2 representing expansion factors
of y-scale limits. The first value is the expansion factor to
the top while the second value is the expansion factor to the bottom.
Values of 0 (the default) represent no expansion relative to the
values in x-scale.}
}
\value{
NULL, invisibly. This function is called for its plotting
  side-effects.
}
\description{
This function takes a data.frame of formatted GWAS results and plots
the p-values in a given interval.
}
\examples{
NULL
}
