% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvalue_tx_plot}
\alias{pvalue_tx_plot}
\title{Plotting p-values along transcript model using grid functions}
\usage{
pvalue_tx_plot(gwas_results, gene_name, genes, transcripts, exons, cds)
}
\arguments{
\item{gene_name}{To complete}

\item{gwas_result}{To complete}

\item{txdb}{To complete}
}
\value{
To complete
}
\description{
This function plots the p-values observed along the genome in
the top panel and a model of the transcripts for genes in the
region in the bottom panel.
}
\examples{
NULL
}
