% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvalue_tx_plot}
\alias{pvalue_tx_plot}
\title{Plotting p-values along transcript model using grid functions}
\usage{
pvalue_tx_plot(
  gwas_results,
  genes,
  transcripts,
  exons,
  cds,
  xscale,
  feature = NULL,
  xexpand = c(0, 0),
  transcript_margins = c(0, 4.1, 0, 2.1),
  pvalue_margins = c(5.1, 4.1, 4.1, 2.1),
  yexpand = c(0, 0)
)
}
\arguments{
\item{gwas_results}{A GRanges object or data.frame of GWAS results.
If a data.frame, must include the columns manhattan_chrom and
manhattan_cpos for coercion to a GRanges. The metadata column
log10p must be present as it will be used to plot the
p-values on the y-axis.}

\item{genes}{A GRanges object containing the known genes for the
reference genome being studied, as obtained from processing
a TxDb object of the GenomicFeatures package.}

\item{transcripts}{A GRangesList object containing the possible
transcripts for the set of genes in the reference genome being used.
Each element of the GRangesList object is a GRanges object containing
the known transcripts for a particular gene, with the name of the
list element being the name of the gene.}

\item{exons}{A GRangesList object containing the documented exons
in the reference genome being used. Each element of the list contains
the exons for a particular transcript, with the name of the list
element being the name of the transcript.}

\item{cds}{A GRangesList object containing the documented coding sequences
in the reference genome being used. Each element of the list contains
the coding sequences for a particular transcript, with the name of the list
element being the name of the transcript.}

\item{xscale}{A GRanges object used to restrict the plotting region.
Typically, it will be a GWAS signal or a gene known to be involved
in the phenotype studied.}

\item{feature}{A GRanges object indicating the position of a feature of
interest to mark with vertical dotted lines. Both the start and end of
the feature will be indicated with a line. If NULL (default), then no
lines are drawn.}

\item{xexpand}{A numeric of length 2 representing expansion factors
of x-scale limits. The first value is the expansion factor to
the left while the second value is the expansion factor to the right.
Values of 0 (the default) represent no expansion relative to the
values in x-scale. Limits will be rounded to the nearest integer
because they need to be represented as a GRanges object.}

\item{transcript_margins}{A numeric of length 4 or that can be recycled to that
length. The margins used by the function \code{\link[grid]{plotViewport}}
to set the margins around the plotting region. By default it is set
to c(0, 4.1, 0, 2.1).}

\item{pvalue_margins}{A numeric of length 4 or that can be recycled to that
length. The margins used by the function \code{\link[grid]{plotViewport}}
to set the margins around the plotting region. By default it is set
to c(5.1, 4.1, 4.1, 2.1).}

\item{yexpand}{A numeric of length 2 representing expansion factors
of y-scale limits. The first value is the expansion factor to
the top while the second value is the expansion factor to the bottom.
Values of 0 (the default) represent no expansion relative to the
values in x-scale.}
}
\value{
NULL, invisibly. The function is called for its side-effect of
  plotting.
}
\description{
This function plots the p-values observed along the genome in
the bottom panel and a model of the transcripts for genes in the
region in the top panel.
}
\details{
The margins of the transcript and p-value plots are set separately
but the left and right margins of each must align in order for the
x-axis values to match on both plots. The function will fail if
this condition is not respected.
}
\examples{
NULL
}
