% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvalue_tx_plot}
\alias{pvalue_tx_plot}
\title{Plotting p-values along transcript model using grid functions}
\usage{
pvalue_tx_plot(
  gwas_results,
  gene_name,
  genes,
  transcripts,
  exons,
  cds,
  transcript_margins = c(0, 4.1, 0, 2.1),
  pvalue_margins = c(5.1, 4.1, 4.1, 2.1)
)
}
\arguments{
\item{gwas_results}{A GRanges object or data.frame of GWAS results.
If a data.frame, must include the columns manhattan_chrom and
manhattan_cpos for coercion to a GRanges. The metadata column
manhattan_log10p must be present as it will be used to plot the
p-values on the y-axis.}

\item{gene_name}{A character. The name of the gene whose transcripts
are to be plotted.}

\item{genes}{A GRanges object containing the known genes for the
reference genome being studied, as obtained from processing
a TxDb object of the GenomicFeatures package.}

\item{transcripts}{A GRangesList object containing the possible
transcripts for the set of genes in the reference genome being used.
Each element of the GRangesList object is a GRanges object containing
the known transcripts for a particular gene, with the name of the
list element being the name of the gene.}

\item{exons}{A GRangesList object containing the documented exons
in the reference genome being used. Each element of the list contains
the exons for a particular transcript, with the name of the list
element being the name of the transcript.}

\item{cds}{A GRangesList object containing the documented coding sequences
in the reference genome being used. Each element of the list contains
the coding sequences for a particular transcript, with the name of the list
element being the name of the transcript.}

\item{transcript_margins}{A numeric of length 4 or that can be recycled to that
length. The margins used by the function \code{\link[grid]{plotViewport}}
to set the margins around the plotting region. By default it is set
to c(0, 4.1, 0, 2.1).}

\item{pvalue_margins}{A numeric of length 4 or that can be recycled to that
length. The margins used by the function \code{\link[grid]{plotViewport}}
to set the margins around the plotting region. By default it is set
to c(5.1, 4.1, 4.1, 2.1).}
}
\value{
NULL, invisibly. The function is called for its side-effect of
  plotting.
}
\description{
This function plots the p-values observed along the genome in
the bottom panel and a model of the transcripts for genes in the
region in the top panel.
}
\details{
The margins of the transcript and p-value plots are set separately
but the left and right margins of each must align in order for the
x-axis values to match on both plots. The function will fail if
this condition is not respected.
}
\examples{
NULL
}
