% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvalueGrob}
\alias{pvalueGrob}
\title{A grob representing p-values of a GWAS analysis at a locus to plot using grid functions}
\usage{
pvalueGrob(
  gwas_results,
  interval,
  feature = NULL,
  yexpand = c(0, 0),
  merging_gap = 10^9,
  cex.points = 0.5
)
}
\arguments{
\item{gwas_results}{A GRanges object or data.frame of GWAS results.
If a data.frame, must include the columns manhattan_chrom and
manhattan_cpos for coercion to a GRanges. The metadata column
log10p must be present as it will be used to plot the
p-values on the y-axis.}

\item{interval}{A GRanges object used to subset the plotting to a given
region.}

\item{feature}{A GRanges object indicating the position of a feature of
interest to mark with vertical dotted lines. Both the start and end of
the feature will be indicated with a line. If NULL (default), then no
lines are drawn.}

\item{yexpand}{A numeric of length 2 representing expansion factors
of y-scale limits. The first value is the expansion factor to
the top while the second value is the expansion factor to the bottom.
Values of 0 (the default) represent no expansion relative to the
values in x-scale.}

\item{merging_gap}{A numeric value specifying how far apart various
signals in the interval GRanges object can be in order to be merged.
This should be a very high value (default = 10^10) such that all ranges
located on a given chromosome will be merged.}

\item{cex.points}{A numeric value. The expansion factor for the points.}
}
\value{
A gTree object that can be plotted with grid.draw() to display the p-values.
}
\description{
This function takes a data.frame of formatted GWAS results and returns
a gTree object representing a p-value plot in a given interval.
}
\examples{
NULL
}
