% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_plot.R
\name{read_kmer_pvalues}
\alias{read_kmer_pvalues}
\title{A function that reads the (sorted) k-mer p-values from the output of a k-mer GWAS analysis}
\usage{
read_kmer_pvalues(kmer_file, max_kmers, kmer_length)
}
\arguments{
\item{kmer_file}{To complete}

\item{max_kmers}{To complete}

\item{kmer_length}{To complete}
}
\value{
To complete
}
\description{
Details
}
\examples{
NULL
}
