% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{transcriptsGrob}
\alias{transcriptsGrob}
\title{Generate a grob of all the possible transcripts for genes in a genomic region using grid}
\usage{
transcriptsGrob(
  genes,
  transcripts,
  exons,
  cds,
  xscale,
  highlight = NULL,
  first_tx_only = FALSE,
  draw_arrows = FALSE
)
}
\arguments{
\item{genes}{A GRanges object containing the known genes for the
reference genome being studied, as obtained from processing
a TxDb object of the GenomicFeatures package.}

\item{transcripts}{A GRangesList object containing the possible
transcripts for the set of genes in the reference genome being used.
Each element of the GRangesList object is a GRanges object containing
the known transcripts for a particular gene, with the name of the
list element being the name of the gene.}

\item{exons}{A GRangesList object containing the documented exons
in the reference genome being used. Each element of the list contains
the exons for a particular transcript, with the name of the list
element being the name of the transcript.}

\item{cds}{A GRangesList object containing the documented coding sequences
in the reference genome being used. Each element of the list contains
the coding sequences for a particular transcript, with the name of the list
element being the name of the transcript.}

\item{xscale}{A GRanges object used to set the limits of the x-axis
and to subset the genes object such that only the ones located in the 
interval are plotted.}

\item{highlight}{A GRanges object indicating a region to highlight
with a red-coloured rectangle to draw attention on a particular
region. If \code{NULL} (the default), then no highlight is plotted.}

\item{first_tx_only}{A logical. Whether to plot only the first transcript
of a given gene (default: FALSE).}

\item{draw_arrows}{A logical indicating whether arrows should be drawn
in the middle of condig sequence regions to indicate the direction
of transcription.}
}
\value{
A gTree with the proper viewports and grobs set for plotting all
  transcripts in the provided genomic region.
}
\description{
This function calls transcriptGrob on all individual transcripts
of the genes located in a given genomic region and organizes each separately
in a row of a viewport grid. The object returned is a gTree that can be plotted
with grid.draw().
}
\details{
This function requires as input the full set of genes, transcripts,
exons and coding sequences for the reference genome being used. It 
will check that transcripts exist for all genes provided. This
behaviour may be modified in the future depending on use cases.

It is also an error in the current implementation to try and plot
over an interval that does not overlap any genes.

The function itself does not add any grobs by itself
but instead arranges the viewport layout and calls the
\code{\link{transcriptGrob}} to add the transcripts one by one.
The latter is the function that should be modified for changing
the way that plotted transcripts look.

The number of rows in the plotting viewport is the number of trancripts
for the gene with the most transcripts over the interval. This is done
so that enough space is provided to plot all transcripts for a given
gene below each other.
}
\examples{
NULL
}
